# Optimize `Substs::super_fold_with`.