--- a/src/librustc/ty/subst.rs
+++ b/src/librustc/ty/subst.rs
@@ -306,8 +306,8 @@ impl<'tcx> TypeFoldable<'tcx> for &'tcx Substs<'tcx> {
     fn super_fold_with<'gcx: 'tcx, F: TypeFolder<'gcx, 'tcx>>(&self, folder: &mut F) -> Self {
         let params: Vec<_> = self.iter().map(|k| k.fold_with(folder)).collect();
 
-        // If folding doesn't change the substs, it's faster to avoid calling
-        // `mk_substs` and instead reuse the existing substs.
+        // If folding doesn't change the substs, it's faster to avoid
+        // calling `mk_substs` and instead reuse the existing substs.
         if params[..] == self[..] {
             self
         } else {