@eddyb 2016-10-12
So this is an interesting take. I'd like to do it *after* moving to on-stack arrays and slice arena allocations instead of always allocating a `Vec`. Then it could either be an extension of that algorithm or maybe it wouldn't even have a measurable impact etc