@eddyb 2016-10-12
Oh and this doesn't handle `ty::relate` which may have a larger impact.