@Mark-Simulacrum 2016-10-18
Note that it will be at least a few days, more likely a week until I can resume work on the `Substs` improvements, so I'm not sure removing things now is a good idea. I haven't followed this PR closely though, not sure how large/significant the changes here are.