@alexcrichton 2016-10-18
@bors: retry force clean

* restarted buildbot