@eddyb 2016-10-19
@bors r-